The Hussle

+++ Libft/Libft-01/ft_bzero.c
@@ -12,7 +12,7 @@
 
 #include <string.h>
 
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*ptr;
